+++ model/src/commonMain/kotlin/com/firsttimeinforever/intellij/pdf/viewer/BrowserMessages.kt
@@ -6,6 +6,9 @@ import kotlinx.serialization.Serializable
 
 object BrowserMessages {
   @Serializable
+  class BrowserReady
+
+  @Serializable
   data class InitialViewProperties(val properties: ViewProperties)
 
   @Serializable