+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Wait for browser to load before sending messages
 - Add retries to forward search
 
 ## 0.18.2